chore(onboarding): drop V2 suffix — rename folder + files
Cosmetic cleanup. Now that there's no legacy folder to clash with, the
v2 onboarding can take the canonical name.

Folder:    src/features/onboarding-v2 → src/features/onboarding
File renames inside:
  OnboardingV2.jsx              → Onboarding.jsx
  onboarding-v2.css             → onboarding.css
  components/ProgressV2.jsx     → components/Progress.jsx
  steps/GenresStepV2.jsx        → steps/GenresStep.jsx
  steps/MoviesStepV2.jsx        → steps/MoviesStep.jsx
  steps/RatingStepV2.jsx        → steps/RatingStep.jsx

Identifier renames inside files (function/component names, comments,
CSS class .onboarding-v2 → .onboarding) and updated the cross-imports
that referenced the old names.

Router: lazy import and route component reference updated; dropped the
/onboarding-v2 alias (was rollout scaffolding, now redundant).

Service comments in shared/services/onboarding.js updated to reference
'Onboarding' instead of 'OnboardingV2'.

Intentionally NOT renamed:
  - localStorage key 'ff_onboarding_v2_moods' (export MOODS_LS_KEY) —
    keeping the value preserves any persisted moods for users who
    started onboarding before the rename. Comment added.
